Rename files in ff dir.
