+++ src/data/ff/bondTerm.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2023 Team Dissolve and contributors
 
-#include "data/ff/bondterm.h"
-#include "data/ff/atomtype.h"
+#include "data/ff/bondTerm.h"
+#include "data/ff/atomType.h"
 #include "data/ff/ff.h"
 
 ForcefieldBondTerm::ForcefieldBondTerm(std::string_view typeI, std::string_view typeJ, BondFunctions::Form form,
