+++ src/neta/ringAtom.cpp
@@ -3,7 +3,7 @@
 
 #include "neta/ringAtom.h"
 #include "classes/speciesAtom.h"
-#include "data/ff/atomtype.h"
+#include "data/ff/atomType.h"
 #include <algorithm>
 #include <utility>
